# OpenApiNetworkLayer/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/workspace/update":{"put":{"tags":["workspace-controller"],"operationId":"updateWorkspace","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWorkspaceEvent"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StringMessageResponse"}}}}}}},"/admin/authorities/add":{"put":{"tags":["admin-controller"],"operationId":"makeUserAdmin","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StringMessageResponse"}}}}}}},"/admin + /email/template/update":{"put":{"tags":["email-controller"],"operationId":"updateEmailTemplate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailTemplate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmailTemplate"}}}}}}},"/workspace/removeMember":{"post":{"tags":["workspace-controller"],"operationId":"removeMemberFromWorkspace","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RemoveMemberFromWorkspaceEvent"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StringMessageResponse"}}}}}}},"/workspace/create":{"post":{"tags":["workspace-controller"],"operationId":"createWorkspace","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWorkspaceEvent"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StringMessageResponse"}}}}}}},"/workspace/addMember":{"post":{"tags":["workspace-controller"],"operationId":"addMemberToWorkspace","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddMemberToWorkspaceEvent"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StringMessageResponse"}}}}}}},"/auth/signup":{"post":{"tags":["authentication-controller"],"operationId":"signUp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StringMessageResponse"}}}}}}},"/auth/signin":{"post":{"tags":["authentication-controller"],"operationId":"signIn","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignInRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponseUserSignInResponse"}}}}}}},"/auth/password/reset":{"post":{"tags":["password-reset-controller"],"operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StringMessageResponse"}}}}}}},"/auth/password/request-reset":{"post":{"tags":["password-reset-controller"],"operationId":"requestPasswordReset","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StringMessageResponse"}}}}}}},"/admin + /email/template/create":{"post":{"tags":["email-controller"],"operationId":"createEmailTemplate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailTemplate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmailTemplate"}}}}}}},"/workspaces/members":{"get":{"tags":["workspaces-controller"],"operationId":"getAllMembersOfWorkspace","parameters":[{"name":"workspaceId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/workspaces/getById":{"get":{"tags":["workspaces-controller"],"operationId":"getWorkspaceById","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Workspace"}}}}}}},"/workspaces/all":{"get":{"tags":["workspaces-controller"],"operationId":"getAllWorkspaces","parameters":[{"name":"page","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Workspace"}}}}}}}},"/user/getMyInfo":{"get":{"tags":["user-controller"],"operationId":"getMyInfo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserInformation"}}}}}}},"/auth/verify-email":{"get":{"tags":["email-verification-controller"],"operationId":"verifyEmail","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/auth/refresh-token":{"get":{"tags":["authentication-controller"],"operationId":"refreshAccessToken","parameters":[{"name":"refresh-token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponseRefreshTokenResponse"}}}}}}},"/admin/user/get":{"get":{"tags":["admin-controller"],"operationId":"getUser","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponseUser"}}}}}}},"/admin/getAllUsers":{"get":{"tags":["admin-controller"],"operationId":"getAllUsers","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponseListUser"}}}}}}},"/admin + /email/template/getById":{"get":{"tags":["email-controller"],"operationId":"getEmailTemplateById","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmailTemplate"}}}}}}},"/admin + /email/template/getAll":{"get":{"tags":["email-controller"],"operationId":"getAllEmailTemplates","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmailTemplate"}}}}}}}},"/workspace/delete":{"delete":{"tags":["workspace-controller"],"operationId":"deleteWorkspace","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteWorkspaceEvent"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StringMessageResponse"}}}}}}},"/admin/user/delete":{"delete":{"tags":["admin-controller"],"operationId":"deleteUser","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StringMessageResponse"}}}}}}},"/admin/authorities/remove":{"delete":{"tags":["admin-controller"],"operationId":"removeUserAdmin","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StringMessageResponse"}}}}}}},"/admin + /email/template/delete":{"delete":{"tags":["email-controller"],"operationId":"deleteEmailTemplate","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UpdateWorkspaceEvent":{"required":["description","id","name"],"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"ownerId":{"type":"string"}}},"StringMessageResponse":{"type":"object","properties":{"message":{"type":"string"}}},"EmailTemplate":{"required":["content","name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"content":{"type":"string"}}},"RemoveMemberFromWorkspaceEvent":{"required":["userId","workspaceId"],"type":"object","properties":{"workspaceId":{"type":"string"},"userId":{"type":"string"},"ownerId":{"type":"string"}}},"CreateWorkspaceEvent":{"required":["description","name"],"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"ownerId":{"type":"string"}}},"AddMemberToWorkspaceEvent":{"required":["userId","workspaceId"],"type":"object","properties":{"workspaceId":{"type":"string"},"userId":{"type":"string"},"ownerId":{"type":"string"}}},"CreateUserRequest":{"required":["email","fullName","nationID","password","phoneNumber"],"type":"object","properties":{"fullName":{"maxLength":100,"minLength":2,"type":"string"},"phoneNumber":{"pattern":"^01[0125][0-9]{8}$","type":"string"},"password":{"maxLength":20,"minLength":6,"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&*(),.?\":{}|<>]).{8,}$","type":"string"},"email":{"type":"string"},"nationID":{"pattern":"^[0-9]{14}$","type":"string"}}},"UserSignInRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":20,"minLength":6,"type":"string"}}},"MessageResponseUserSignInResponse":{"type":"object","properties":{"message":{"type":"string"},"data":{"$ref":"#/components/schemas/UserSignInResponse"}}},"UserSignInResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"fullName":{"type":"string"},"token":{"type":"string"},"refreshToken":{"type":"string"}}},"ResetPasswordRequest":{"required":["email","newPassword","oobCode"],"type":"object","properties":{"email":{"type":"string"},"oobCode":{"type":"string"},"newPassword":{"maxLength":20,"minLength":6,"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&*(),.?\":{}|<>]).{8,}$","type":"string"}}},"Workspace":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"ownerId":{"type":"string"}}},"UserInformation":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"fullName":{"type":"string"},"authorities":{"type":"array","items":{"type":"string"}},"nationID":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"phone":{"type":"string"},"phoneVerified":{"type":"boolean"},"emailVerified":{"type":"boolean"}}},"MessageResponseRefreshTokenResponse":{"type":"object","properties":{"message":{"type":"string"},"data":{"$ref":"#/components/schemas/RefreshTokenResponse"}}},"RefreshTokenResponse":{"type":"object","properties":{"newJwt":{"type":"string"},"newRefreshToken":{"type":"string"}}},"MessageResponseUser":{"type":"object","properties":{"message":{"type":"string"},"data":{"$ref":"#/components/schemas/User"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"fullName":{"type":"string"},"authorities":{"type":"array","items":{"type":"string"}},"password":{"type":"string","writeOnly":true},"nationID":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"phone":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"lastModifiedDate":{"type":"string","format":"date-time"},"phoneVerified":{"type":"boolean"},"emailVerified":{"type":"boolean"}}},"MessageResponseListUser":{"type":"object","properties":{"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}},"DeleteWorkspaceEvent":{"required":["id"],"type":"object","properties":{"id":{"type":"string"},"ownerId":{"type":"string"}}}}}}